Move task class into the master branch
